src/3. Retrieval Augmented Generation/pyproject.toml
[project]
name = "RAG_primitves"
version = "0.1.0"
description = "A set of RAG primitives, for independend execution"
authors = ["Andrzej Wodecki <andrzej.wodecki@example.com>"]
requires-python = ">=3.9"
dependencies = [
    "langchain==0.3.24",
    "langchain-community==0.3.22",
    "langchain-core==0.3.55",
    "langchain-text-splitters==0.3.8",
    "langchain-openai==0.3.11",
    "langchain-chroma==0.2.3",
    "langchain-experimental>=0.3.0",
    "unstructured[pdf]>=0.17.2",
    "libmagic>=1.0",
    "python-magic>=0.4.27",
    "faiss-cpu>=1.12.0",
    "python-dotenv>=1.1.0",
    "pypdf>=5.4.0",
    "reportlab>=4.4.4",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.4",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=5.1.1",
    "ragas>=0.3.5",
    "neo4j>=5.28.2",
    "chromadb>=0.6.3",
    "langchain-neo4j>=0.5.0",
    "pydantic>=2.11.4",
    "faker>=37.8.0",
    "markdown>=3.9",
    "weasyprint>=66.0",
    "toml>=0.10.2",
]
